Calculating commit diff
- Created function get_diff that calculates SAFA delta between starting and ending commit
- Added repo diff functionality to create project tool

+++ publish.sh
@@ -1,1 +1,1 @@
-pytest && echo "Tests Are Passing!" && python3 -m build && python3 -m twine upload --repository pypi dist/* && rm -rf dist
+mypy && pytest && echo "Tests Are Passing!" && python3 -m build && python3 -m twine upload --repository pypi dist/* && rm -rf dist
